myListページを削除
