+++ lib/ui/pages/album_list_page/src/alert_dialog.dart
@@ -2,10 +2,11 @@ import 'package:flutter/material.dart';
 
 class DisplayDialog extends StatelessWidget {
   const DisplayDialog({
+    Key? key,
     required this.title,
     required this.content,
     required this.onPressed,
-  });
+  }) : super(key: key);
 
   final String title;
   final String content;
