--- lib/ui/pages/my_list_page/my_list_page.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:my_collection/themes/app_colors.dart';
-import 'package:my_collection/ui/components/components.dart';
-import 'package:my_collection/ui/pages/my_list_page/src/my_list_page_body.dart';
-
-class MyListPage extends ConsumerWidget {
-  const MyListPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Scaffold(
-      backgroundColor: AppColors.scaffoldColor,
-      appBar: AppBar(
-        backgroundColor: AppColors.barColor,
-        title: const AppBarText('My思い出一覧'),
-      ),
-      body: const MyListPgeBody(),
-    );
-  }
-}
