+++ pubspec.yaml
@@ -9,31 +9,32 @@ version: 1.3.0+103000
 # 1.2.3 => 102030
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: '3.3.1'
 
 dependencies:
   flutter:
     sdk: flutter
 
   cached_network_image: ^3.0.0
-  cloud_firestore: ^3.1.1
+  cloud_firestore: ^3.5.1
   cupertino_icons: ^1.0.4
   exif: ^3.1.1
   firebase_analytics: ^8.3.4
   firebase_auth: ^3.3.7
-  firebase_core: ^1.10.2
-  firebase_storage: ^10.2.4
+  firebase_core: ^1.24.0
+  firebase_storage: ^10.3.11
   fluttertoast: ^8.0.8
   flutter_hooks: ^0.18.0
   flutter_native_splash: ^2.0.5
   flutter_slidable: ^0.6.0
   flutter_svg: ^0.22.0
   freezed_annotation: ^1.1.0
-  google_sign_in: ^5.2.3
-  google_maps_flutter: ^2.1.1
+  google_sign_in: ^5.4.2
+  google_maps_flutter: ^2.2.1
   hooks_riverpod: ^1.0.0-dev.7
   image_cropper: ^1.5.0
-  image_picker: ^0.8.4+9
+  image_picker: ^0.8.6
   introduction_screen: ^3.0.0
   json_serializable: ^6.1.1
   location: ^4.3.0
